// petstore/servers/node/tsp-output/@typespec/http-server-javascript/models/all/pet-store.ts
export interface Owners<Context = unknown> {
  /**
   * Gets an instance of the resource.
   */
  get(ctx: Context, ownerId: bigint): Promise<Owner | PetStoreError>;

  /**
   * Updates an existing instance of the resource.
   */
  update(
    ctx: Context,
    ownerId: bigint,
    properties: OwnerUpdate,
  ): Promise<Owner | PetStoreError>;

  /**
   * Deletes an existing instance of the resource.
   */
  delete(
    ctx: Context,
    ownerId: bigint,
  ): Promise<ResourceDeletedResponse | PetStoreError>;

  /**
   * Creates a new instance of the resource.
   */
  create(
    ctx: Context,
    resource: OwnerCreate,
  ): Promise<Owner | OwnerResourceCreatedResponse | PetStoreError>;

  /**
   * Lists all instances of the resource.
   */
  list(ctx: Context): Promise<OwnerCollectionWithNextLink | PetStoreError>;
}
// ...
export interface PetStoreError {
  code: number;

  message: string;
}

export interface Pet {
  id: number;

  name: string;

  tag?: string;

  age: number;

  ownerId: bigint;
}

export interface Toy {
  id: bigint;

  petId: bigint;

  name: string;
}

export interface Owner {
  id: bigint;

  name: string;

  age: number;
}